# docs/openapi.yaml
openapi: 3.1.0
info:
  title: BoardGameTools API
  description: |
    REST API fuer BoardGameTools — Brettspiel-Verwaltung, Sessions, Events mit Voting, Gruppen und Spielereihen.

    ## Authentifizierung
    Die meisten Endpunkte erfordern eine gueltige NextAuth.js Session (Cookie-basiert).
    Oeffentliche Endpunkte unter `/api/public/` und `/api/auth/` sind ohne Login erreichbar.

    ## Fehlerformat
    Alle Fehler folgen dem Format: `{ "error": "Deutsche Fehlermeldung" }`

    ## Rate Limiting
    Oeffentliche Endpunkte sind per IP rate-limited (Upstash Redis + In-Memory Fallback).
  version: 0.40.1
  contact:
    name: BoardGameTools
  license:
    name: MIT
servers:
  - url: http://localhost:3000
    description: Lokale Entwicklung
  - url: https://boardgametools.vercel.app
    description: Production
...
components:
  schemas:
...
    Group:
      type: object
      properties:
        id: { type: string, format: cuid }
        name: { type: string }
        description: { type: string, nullable: true }
        ownerId: { type: string }
        shareToken: { type: string, nullable: true }
        hasPassword: { type: boolean }
...
  securitySchemes:
    session:
      type: apiKey
      in: cookie
      name: next-auth.session-token
      description: NextAuth.js Session Cookie
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: API Access Token aus `POST /api/mobile/v1/auth/login` oder `refresh`
...
security:
  - session: []
paths:
...
  /api/mobile/v1/groups:
    get:
      tags: [Mobile]
      summary: Gruppen-Liste
      security: [ bearerAuth: [] ]
      responses:
        '200': { description: Gruppen-Array }
        '401': { $ref: '#/components/responses/Unauthorized' }
    post:
      tags: [Mobile]
      summary: Gruppe anlegen
      security: [ bearerAuth: [] ]
      requestBody:
        required: true
        content:
          application/json: { schema: { $ref: '#/components/schemas/Group' } }
      responses:
        '201': { description: Angelegte Gruppe }
        '401': { $ref: '#/components/responses/Unauthorized' }
...
  /api/mobile/v1/groups/{id}/polls:
    get:
      tags: [Mobile]
      summary: Gruppen-Umfragen
      security: [ bearerAuth: [] ]
      parameters: [{ $ref: '#/components/parameters/IdParam' }]
      responses:
        '200': { description: Umfragen-Array }
        '401': { $ref: '#/components/responses/Unauthorized' }
    post:
      tags: [Mobile]
      summary: Umfrage erstellen
      security: [ bearerAuth: [] ]
      parameters: [{ $ref: '#/components/parameters/IdParam' }]
      requestBody:
        required: true
        content:
          application/json: { schema: { $ref: '#/components/schemas/GroupPoll' } }
      responses:
        '201': { description: Umfrage erstellt }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /api/mobile/v1/groups/{id}/polls/{pollId}/vote:
    post:
      tags: [Mobile]
      summary: Umfrage-Option waehlen
      security: [ bearerAuth: [] ]
      parameters:
        - { $ref: '#/components/parameters/IdParam' }
        - name: pollId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                optionId: { type: string }
      responses:
        '200': { description: Stimme gespeichert }
        '401': { $ref: '#/components/responses/Unauthorized' }
...
  /api/mobile/v1/uploads:
    post:
      tags: [Upload]
      summary: Bild-Upload fuer Mobile
      security: [ bearerAuth: [] ]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                image:
                  type: string
                  format: binary
      responses:
        '201': { description: Upload-URL und Pfad }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
...